requirements.txt
arabic-reshaper
asgiref
asn1crypto
certifi
cffi
chardet
click
colorama
cryptography
cssselect2
dj-database-url
Django
django-autoslug
django-ckeditor
django-crispy-forms
django-heroku
django-js-asset
django-widget-tweaks
future
gunicorn
html5lib
idna
lxml
oscrypto
Pillow
psycopg2
psycopg2-binary
pycparser
pyHanko
pyhanko-certvalidator
PyJWT
pypdf
PyPDF2
pypng
python-bidi
pytz
PyYAML
qrcode
razorpay
reportlab
requests
six
sqlparse
svglib
tinycss2
twilio
typing_extensions
tzdata
tzlocal
uritools
urllib3
webencodings
whitenoise
xhtml2pdf
google-generativeai